+++ default.py
@@ -190,6 +190,8 @@ def main():
         show_localized_notification(kodi_helper, 32000, 32007)
     except UnableGetStreamListException:
         show_localized_notification(kodi_helper, 32000, 32010)
+    except StrangeXBMCException:
+        show_localized_notification(kodi_helper, 32000, 32011)
     except StreamHasNotStarted:
         show_localized_notification(kodi_helper, 30004, 30008, xbmcgui.NOTIFICATION_INFO)
     except TipsportMsg as e: