+++ resources/lib/kodi_helper.py
@@ -5,7 +5,8 @@ from urllib.parse import parse_qs, urlencode
 import xbmc
 import xbmcaddon
 import xbmcvfs
-from .tipsport_stream_generator import Quality, Site
+from .tipsport_exceptions import StrangeXBMCException
+from .tipsport_stream_generator import Site
 from .user_data import UserData
 from .utils import log
 try:
@@ -82,7 +83,7 @@ class KodiHelper:
         if not xbmcvfs.exists(self.tmp_path):
             xbmcvfs.mkdirs(self.tmp_path)
         self.version = addon.getAddonInfo('version')
-        self.user_data = UserData(addon.getSetting('username'), addon.getSetting('password'), self.__get_quality(addon), self.__get_site(addon))
+        self.user_data = UserData(addon.getSetting('username'), addon.getSetting('password'), self.__get_site(addon))
         self.send_crash_reports = addon.getSetting('send_crash_reports') == 'true'
         self.show_all_matches = addon.getSetting('show_all_matches') == 'true'
         self.icon = addon.getAddonInfo('icon')
@@ -93,16 +94,15 @@ class KodiHelper:
         return CAN_GENERATE_LOGOS and self.can_generate_logos_settings
 
     @staticmethod
-    def __get_quality(addon):
-        return Quality.parse(addon.getSetting('quality'))
-
-    @staticmethod
     def __get_site(addon):
         return Site.parse(addon.getSetting('site'))
 
     @staticmethod
     def get_addon():
-        return xbmcaddon.Addon()
+        try:
+            return xbmcaddon.Addon()
+        except:
+            raise StrangeXBMCException()
 
     def build_url(self, query):
         return self.base_url + '?' + urlencode(query)