+++ resources/lib/stream_strategy.py
@@ -5,7 +5,6 @@ try:
 except:
     from urllib.parse import urljoin
 from .utils import log
-from .quality import Quality
 from .stream import RTMPStream, PlainStream
 from .tipsport_exceptions import UnableGetStreamMetadataException
 
@@ -14,7 +13,7 @@ class RTMPStreamStrategy:
     def __init__(self, url):
         self._url = url
 
-    def get_stream(self, _):
+    def get_stream(self):
         try:
             playpath = (self._url.split('/'))[-1]
             url = self._url.replace('/' + playpath, '')
@@ -30,15 +29,14 @@ class UrlImgStreamStrategy:
         self._url = url
         self._session = session
 
-    def get_stream(self, quality):
+    def get_stream(self):
         hls_data = self._session.get(self._url)
         data = json.loads(hls_data.content)
         attribute_name = 'hlsUrl'
         if attribute_name not in data or data[attribute_name] is None:
             return None
         m3u8_playlist_url = data[attribute_name]
-        m3u8_playlist = self._session.get(m3u8_playlist_url)
-        stream_url = _get_stream_url_from_m3u8(m3u8_playlist, m3u8_playlist_url, quality)
+        stream_url = m3u8_playlist_url
         return PlainStream(stream_url)
 
 
@@ -47,10 +45,9 @@ class HLSStreamStrategy:
         self._url = url
         self._session = session
 
-    def get_stream(self, quality):
+    def get_stream(self):
         url_query = _get_query_part(self._url)
-        m3u8_playlist = self._session.get(self._url)
-        stream_url = _get_stream_url_from_m3u8(m3u8_playlist, self._url, quality)
+        stream_url = self._url
         if url_query:
             if '?' in stream_url:
                 stream_url = stream_url.strip() + '&' + url_query
@@ -65,14 +62,13 @@ class UrlPerformeStreamStrategy:
         self._url = url
         self._session = session
 
-    def get_stream(self, quality):
+    def get_stream(self):
         xml_data = self._session.get(self._url)
         possible_m3u8_urls = re.findall('<streamLaunchCode><!\[CDATA\[(.*)\]\]></streamLaunchCode>', xml_data.text.replace('\n', ''))
         possible_m3u8_urls = [m3u8_url for m3u8_url in possible_m3u8_urls if 'm3u8' in m3u8_url or 'hls' in m3u8_url]
         if len(possible_m3u8_urls) == 0:
             return None
-        m3u8_playlist = self._session.get(possible_m3u8_urls[-1])
-        stream_url = _get_stream_url_from_m3u8(m3u8_playlist, self._url, quality)
+        stream_url = self._url
         return PlainStream(stream_url)
 
 
@@ -81,11 +77,8 @@ class UrlAguraStrategy:
         self._url = url
         self._session = session
 
-    def get_stream(self, quality):
-        json_data_response = self._session.get(self._url)
-        json_data = json.loads(json_data_response.text)
-        m3u8_playlist = self._session.get(json_data['url'])
-        stream_url = _get_stream_url_from_m3u8(m3u8_playlist, self._url, quality)
+    def get_stream(self):
+        stream_url = self._url
         return PlainStream(stream_url)
 
 
@@ -94,11 +87,8 @@ class TvComStreamStrategy:
         self._url = url
         self._session = session
 
-    def get_stream(self, quality):
-        json_data_response = self._session.get(self._url)
-        json_data = json.loads(json_data_response.text)
-        m3u8_playlist = self._session.get(json_data['url']['hls']['url'])
-        stream_url = _get_stream_url_from_m3u8(m3u8_playlist, self._url, quality)
+    def get_stream(self):
+        stream_url = self._url
         return PlainStream(stream_url)
 
 
@@ -106,38 +96,9 @@ class NoneStrategy:
     def __init__(self, relative_url):
         self.relative_url = relative_url
 
-    def get_stream(self, _):
+    def get_stream(self):
         log('NoneStrategy: ' + self.relative_url)
         return None
-
-
-def _select_stream_by_quality(list_of_streams, quality):
-    """List is ordered from the lowest to the best quality"""
-    if len(list_of_streams) == 0:
-        log('list_of_streams is empty')
-        raise UnableGetStreamMetadataException('List of streams by quality is empty')
-    if len(list_of_streams) == quality + 1:
-        return list_of_streams[quality]
-    if quality in [Quality.LOW, Quality.MID]:
-        return list_of_streams[0]
-    return list_of_streams[-1]
-
-
-def _get_stream_url_from_m3u8(m3u8_playlist, m3u8_playlist_url, quality):
-    urls = _get_m3u8_urls_sorted(m3u8_playlist.text)
-    stream_url = _select_stream_by_quality(urls, quality)
-    if _is_stream_relative(stream_url):
-        url = m3u8_playlist_url.split('?')[0]
-        stream_url = urljoin(url, stream_url)
-    return stream_url
-
-
-# Return tuple: (bandwidth, resolution, avg_bandwidth, url)
-M3U8_PATTERN = re.compile('#EXT-X-STREAM-INF:(?:BANDWIDTH=(?P<bandwidth>\d+),?|RESOLUTION=\d+x(?P<resolution>\d+),?|AVERAGE-BANDWIDTH=(?P<avg_bandwidth>\d+),?|[a-zA-Z_-]+=.*?,?)*\n(?P<url>.*)')
-def _get_m3u8_urls_sorted(m3u8_playlist):
-    urls = re.findall(M3U8_PATTERN, m3u8_playlist)
-    urls = sorted(urls, key=lambda x: (x[1], x[2], x[0], -len(x[3])))
-    return [x[3] for x in urls]
 
 
 def _is_stream_relative(stream_url):